refactor: add wrapper function for treemorph rewriter
