+++ crates/conjure_core/src/rule_engine/mod.rs
@@ -1,5 +1,8 @@
 pub use linkme::distributed_slice;
 
+mod rewrite_morph;
+pub use rewrite_morph::rewrite_morph;
+
 /// This procedural macro registers a decorated function with `conjure_rules`' global registry, and
 /// adds the rule to one or more `RuleSet`'s.
 ///
